+++ dflib-echarts/src/test/java/org/dflib/echarts/ScatterSeriesTest.java
@@ -1,0 +1,30 @@
+package org.dflib.echarts;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class ScatterSeriesTest extends GenerateScriptHtmlTest {
+
+    @Test
+    public void itemStyle() {
+
+        ScatterItemStyle style = ScatterItemStyle.of()
+                .color("#ffffff")
+                .borderColor("#eeeeeee")
+                .borderWidth(2)
+                .opacity(0.55);
+
+        String s1 = ECharts.chart().series(SeriesOpts.ofScatter(), "y1").generateScriptHtml("_tid", df2);
+        assertFalse(s1.contains("itemStyle"), s1);
+
+        String s2 = ECharts.chart().series(SeriesOpts.ofScatter().itemStyle(style), "y1").generateScriptHtml("_tid", df2);
+
+        assertTrue(s2.contains("itemStyle"), s2);
+        assertTrue(s2.contains("color: '#ffffff',"), s2);
+        assertTrue(s2.contains("borderColor: '#eeeeeee',"), s2);
+        assertTrue(s2.contains("borderWidth: 2,"), s2);
+        assertTrue(s2.contains("opacity: 0.55"), s2);
+    }
+}