+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #362 Aggregating expressions do not generate proper QL
 * #363 BooleanSeries.firstFalse()
 * #367 ECharts: support "itemStyle" for "bar" series
+* #368 ECharts: support "itemStyle" for "scatter" series
 
 ## Release 1.0.0-RC2